# perf(snowflake): send everything through pyarrow