--- a/ibis/backends/base/__init__.py
+++ b/ibis/backends/base/__init__.py
@@ -504,6 +504,8 @@ class BaseBackend(abc.ABC, _FileIOHandler):
     supports_temporary_tables = False
 
     def __init__(self, *args, **kwargs):
+        from ibis.formats.pandas import PandasConverter
+
         self._con_args: tuple[Any] = args
         self._con_kwargs: dict[str, Any] = kwargs
         # expression cache
@@ -514,6 +516,7 @@ class BaseBackend(abc.ABC, _FileIOHandler):
             generate_name=functools.partial(util.gen_name, "cache"),
             key=lambda expr: expr.op(),
         )
+        self._pandas_converter = PandasConverter
 
     def __getstate__(self):
         return dict(_con_args=self._con_args, _con_kwargs=self._con_kwargs)