--- a/ibis/backends/bigquery/__init__.py
+++ b/ibis/backends/bigquery/__init__.py
@@ -28,7 +28,6 @@ from ibis.backends.bigquery.client import (
 )
 from ibis.backends.bigquery.compiler import BigQueryCompiler
 from ibis.backends.bigquery.datatypes import schema_from_bigquery, schema_to_bigquery
-from ibis.formats.pandas import convert_pandas_dataframe
 
 with contextlib.suppress(ImportError):
     from ibis.backends.bigquery.udf import udf  # noqa: F401
@@ -337,7 +336,7 @@ class Backend(BaseSQLBackend):
     def fetch_from_cursor(self, cursor, schema):
         arrow_t = self._cursor_to_arrow(cursor)
         df = arrow_t.to_pandas(timestamp_as_object=True)
-        return convert_pandas_dataframe(df, schema)
+        return self._pandas_converter.convert_frame(df, schema)
 
     def _cursor_to_arrow(
         self,