--- a/ibis/backends/pandas/core.py
+++ b/ibis/backends/pandas/core.py
@@ -131,7 +131,6 @@ from ibis.backends.pandas.dispatch import (
     pre_execute,
 )
 from ibis.backends.pandas.trace import trace
-from ibis.formats.pandas import convert_pandas_dataframe
 
 integer_types = np.integer, int
 floating_types = (numbers.Real,)
@@ -492,13 +491,15 @@ def execute_and_reset(
 
 
 def _apply_schema(op: ops.Node, result: pd.DataFrame | pd.Series):
+    from ibis.formats.pandas import PandasConverter
+
     assert isinstance(op, ops.Node), type(op)
     if isinstance(result, pd.DataFrame):
         df = result.reset_index().loc[:, list(op.schema.names)]
-        return convert_pandas_dataframe(df, op.schema)
+        return PandasConverter.convert_frame(df, op.schema)
     elif isinstance(result, pd.Series):
         schema = op.to_expr().as_table().schema()
-        df = convert_pandas_dataframe(result.to_frame(), schema)
+        df = PandasConverter.convert_frame(result.to_frame(), schema)
         return df.iloc[:, 0].reset_index(drop=True)
     else:
         return result