--- a/ibis/backends/pyspark/__init__.py
+++ b/ibis/backends/pyspark/__init__.py
@@ -29,7 +29,6 @@ from ibis.backends.pyspark import ddl
 from ibis.backends.pyspark.client import PySparkTable
 from ibis.backends.pyspark.compiler import PySparkExprTranslator
 from ibis.backends.pyspark.datatypes import dtype_from_pyspark
-from ibis.formats.pandas import convert_pandas_dataframe
 
 if TYPE_CHECKING:
     import pandas as pd
@@ -212,7 +211,7 @@ class Backend(BaseSQLBackend):
         table_expr = expr.as_table()
         df = self.compile(table_expr, **kwargs).toPandas()
 
-        result = convert_pandas_dataframe(df, table_expr.schema())
+        result = self._pandas_converter.convert_frame(df, table_expr.schema())
         if isinstance(expr, ir.Table):
             return result
         elif isinstance(expr, ir.Column):