## What does this PR Implement?

This PR removes the pyarrow optionality in the snowflake backend in an effort to

1. Remove the divergence between slow-path-no-pyarrow and fast-path-pyarrow. This simplifies the backend code.
1. Improve performance (numbers below) meaningfully, when pulling back medium to large results.
1. ~Make type handling more consistent (though, it's not clear if that's actually happening in this PR because of the added complexity around post-execute conversion!)~ This is much much simpler after mapping every complex type to `convert_JSON` (see the diff for context).

Some implications:

We're back in the situation of `pyarrow` being limited to version 10 in the lock file.

This doesn't limit end user versions of pyarrow except for snowflake backend
end users. Previously, you could have a snowflake-incompatible version of
pyarrow installed and the snowflake backend would work, but it'd be much slower.

We can artificially install a newer version of pyarrow in our other builds,
but that will slow down CI quite a bit. Perhaps we can add a few special
jobs of the faster-to-test backends that explicitly install `pyarrow>10`

Alternatively, we can risk users having a broken install by leaving the `"pandas"` extras
off, but that seems like a bad idea to me.

## Performance Comparison

### Setup

tpch scale factor 1, lineitem table, ~6MM rows, ran each twice to ensure the warehouse was up and warm

### Baseline, **WITHOUT** PyArrow

```
In [1]: import os

In [2]: from ibis.interactive import *

In [3]: con = ibis.connect(os.environ["SNOWFLAKE_URL"])
li
In [4]: lineitem = con.table("LINEITEM", schema="snowflake_sample_data.tpch_sf1")

In [5]: %time df = lineitem.to_pandas()
CPU times: user 1min 35s, sys: 7.92 s, total: 1min 43s
Wall time: 1min 42s
```

### Candidate, **WITH** PyArrow

```
In [8]: %time df = lineitem.to_pandas()
CPU times: user 9.87 s, sys: 4.17 s, total: 14 s
Wall time: 20.5 s
```

The speedup is pretty not bad!