@cpcloud 2023-06-07
Snowflake backend tests look good:

```
cloud in 🌐 falcon in …/ibis on  snowflake-only-pyarrow is 📦 v5.1.0 via 🐍 v3.10.11 via ❄️  impure (ibis-3.10.11-env)
❯ pytest -m snowflake -n auto --dist loadgroup --snapshot-update
================================================ test session starts =================================================
platform linux -- Python 3.10.11, pytest-7.3.1, pluggy-1.0.0
Using --randomly-seed=3943998521
benchmark: 4.0.0 (defaults: timer=time.perf_counter disable_gc=False min_rounds=5 min_time=0.000005 max_time=1.0 calibration_precision=10 warmup=False warmup_iterations=100000)
rootdir: /home/cloud/ibis
configfile: pyproject.toml
plugins: randomly-3.12.0, hypothesis-6.76.0, profiling-1.7.0, benchmark-4.0.0, clarity-1.0.1, xdist-3.3.1, cov-4.1.0, mock-3.10.0, httpserver-1.0.8, repeat-0.9.1, snapshot-0.9.0
32 workers [959 items]
...xs.x........s......................s........xx...x..............xx.........x.....x...x..xx.x..x...xx.x.x... [ 11%]
x...x..x...x..x.xx...x..x..xx.....x...x...x.........x..x.......x.x.xxxxx.....xx..............x.....x..x......x [ 22%]
.................x.x...........................................x...........x.xx...x.x...x.x.x.x............... [ 34%]
.x........x...x...x..x..x...x..........x..........x...........x.x........xx......x.x.....s...xxxxxxxx.xx...xxx [ 45%]
......xx..x......................x.....xx....x...x........x.xxxx..xxx.....x...x.....x.xx........x............. [ 57%]
..........x.......x......x..x...x............................................................x.........x...... [ 68%]
....x.......................x.........x..................x.....x.............x......x......................... [ 80%]
.........x.....x......................x..x.......x.x.......x.....x.x....x..xx.............x....x..........x... [ 91%]
..xxx....xs...s.x.x...xx.....x....xx..x...x.x..x..x..x.x..x.xx.xxx.xxxxxxxxxxx.                                [100%]
============================== 780 passed, 6 skipped, 173 xfailed in 169.61s (0:02:49) ===============================
```